Add PaintStructJson
