+++ src/openrct2/object/PaintObject.cpp
@@ -202,218 +202,14 @@ void PaintObject::ReadJson(IReadObjectContext* context, json_t& root)
 
             for (const auto& paintStruct : paintStructs)
             {
-                PaintStructDescriptor paint;
-                PaintStructKeyJson keyJson;
-                keyJson.FromJson(paintStruct);
-
-                if (paintStruct.contains("supports"))
-                {
-                    auto supports = paintStruct["supports"];
-                    if (supports.is_string())
-                    {
-                        if (supports == "wooden_a")
-                            paint.Supports = PaintStructDescriptor::SupportsType::WoodenA;
-                    }
-                }
-
-                if (paintStruct.contains("floor"))
-                {
-                    auto floor = paintStruct["floor"];
-                    if (floor.is_string())
-                    {
-                        if (floor == "cork")
-                            paint.Floor = PaintStructDescriptor::FloorType::Cork;
-                    }
-                }
-
-                if (paintStruct.contains("edges"))
-                {
-                    auto edges = paintStruct["edges"];
-                    if (edges.is_string())
-                    {
-                        if (_edgeMappings.find(edges) != _edgeMappings.end())
-                        {
-                            paint.Edges = &_edgeMappings[edges];
-                        }
-                    }
-                }
-
-                if (paintStruct.contains("fences"))
-                {
-                    auto fences = paintStruct["fences"];
-                    if (fences.is_string())
-                    {
-                        if (fences == "ropes")
-                            paint.Fences = PaintStructDescriptor::FenceType::Ropes;
-                    }
-                }
-
-                if (paintStruct.contains("paintType"))
-                {
-                    auto paintType = paintStruct["paintType"];
-                    if (paintType.is_string())
-                    {
-                        if (paintType == "addImageAsParent")
-                            paint.PaintType = PaintStructDescriptor::PaintType::AddImageAsParent;
-                        else if (paintType == "addImageAsChild")
-                            paint.PaintType = PaintStructDescriptor::PaintType::AddImageAsChild;
-                    }
-                }
-
-                if (paintStruct.contains("imageIdBase"))
-                {
-                    auto imageIdBase = paintStruct["imageIdBase"];
-                    if (imageIdBase.is_string())
-                    {
-                        if (imageIdBase == "car0")
-                            paint.ImageIdBase = PaintStructDescriptor::ImageIdBase::Car0;
-                    }
-                }
-
-                if (paintStruct.contains("primaryColour"))
-                {
-                    auto primaryColour = paintStruct["primaryColour"];
-                    if (primaryColour.is_string())
-                    {
-                        if (primaryColour == "vehicleBody")
-                            paint.PrimaryColour = PaintStructDescriptor::Colour::VehicleBody;
-                        else if (primaryColour == "vehicleTrim")
-                            paint.PrimaryColour = PaintStructDescriptor::Colour::VehicleTrim;
-                        else if (primaryColour == "peepTShirt")
-                            paint.PrimaryColour = PaintStructDescriptor::Colour::PeepTShirt;
-                    }
-                }
-
-                if (paintStruct.contains("secondaryColour"))
-                {
-                    auto secondaryColour = paintStruct["secondaryColour"];
-                    if (secondaryColour.is_string())
-                    {
-                        if (secondaryColour == "vehicleBody")
-                            paint.SecondaryColour = PaintStructDescriptor::Colour::VehicleBody;
-                        else if (secondaryColour == "vehicleTrim")
-                            paint.SecondaryColour = PaintStructDescriptor::Colour::VehicleTrim;
-                        else if (secondaryColour == "peepTShirt")
-                            paint.SecondaryColour = PaintStructDescriptor::Colour::PeepTShirt;
-                    }
-                }
-
-                if (paintStruct.contains("primaryColourIndex"))
-                {
-                    auto primaryColourIndex = paintStruct["primaryColourIndex"];
-                    if (primaryColourIndex.is_number())
-                        paint.PrimaryColourIndex = primaryColourIndex;
-                }
-
-                if (paintStruct.contains("secondaryColourIndex"))
-                {
-                    auto secondaryColourIndex = paintStruct["secondaryColourIndex"];
-                    if (secondaryColourIndex.is_number())
-                        paint.SecondaryColourIndex = secondaryColourIndex;
-                }
-
-                if (paintStruct.contains("imageIdOffset"))
-                {
-                    auto imageIdOffset = paintStruct["imageIdOffset"];
-                    if (imageIdOffset.is_string())
-                    {
-                        paint.ImageIdOffset = _imageIdOffsetMapping[imageIdOffset].get();
-                    }
-                }
-
-                if (paintStruct.contains("imageIdOffsetIndex"))
-                {
-                    auto imageIdOffsetIndex = paintStruct["imageIdOffsetIndex"];
-                    if (imageIdOffsetIndex.is_number())
-                        paint.ImageIdOffsetIndex = imageIdOffsetIndex;
-                }
-
-                if (paintStruct.contains("imageIdScheme"))
-                {
-                    auto imageIdScheme = paintStruct["imageIdScheme"];
-                    if (imageIdScheme.is_string())
-                    {
-                        if (imageIdScheme == "misc")
-                            paint.ImageIdScheme = PaintStructDescriptor::Scheme::Misc;
-                    }
-                }
-
-                if (paintStruct.contains("offset_x"))
-                {
-                    auto offsetX = paintStruct["offset_x"];
-                    if (offsetX.is_number())
-                        paint.Offset.x = offsetX;
-                }
-
-                if (paintStruct.contains("offset_y"))
-                {
-                    auto offsetY = paintStruct["offset_y"];
-                    if (offsetY.is_number())
-                        paint.Offset.y = offsetY;
-                }
-
-                if (paintStruct.contains("offset_z"))
-                {
-                    auto offsetZ = paintStruct["offset_z"];
-                    if (offsetZ.is_number())
-                        paint.Offset.z = offsetZ;
-                }
-
-                if (paintStruct.contains("bb_offset_x"))
-                {
-                    auto bbOffsetX = paintStruct["bb_offset_x"];
-                    if (bbOffsetX.is_number())
-                        paint.BoundBox.offset.x = bbOffsetX;
-                }
-
-                if (paintStruct.contains("bb_offset_y"))
-                {
-                    auto bbOffsetY = paintStruct["bb_offset_y"];
-                    if (bbOffsetY.is_number())
-                        paint.BoundBox.offset.y = bbOffsetY;
-                }
-
-                if (paintStruct.contains("bb_offset_z"))
-                {
-                    auto bbOffsetZ = paintStruct["bb_offset_z"];
-                    if (bbOffsetZ.is_number())
-                        paint.BoundBox.offset.z = bbOffsetZ;
-                }
-
-                if (paintStruct.contains("bb_length_x"))
-                {
-                    auto bbLengthX = paintStruct["bb_length_x"];
-                    if (bbLengthX.is_number())
-                        paint.BoundBox.length.x = bbLengthX;
-                }
-
-                if (paintStruct.contains("bb_length_y"))
-                {
-                    auto bbLengthY = paintStruct["bb_length_y"];
-                    if (bbLengthY.is_number())
-                        paint.BoundBox.length.y = bbLengthY;
-                }
-
-                if (paintStruct.contains("bb_length_z"))
-                {
-                    auto bbLengthZ = paintStruct["bb_length_z"];
-                    if (bbLengthZ.is_number())
-                        paint.BoundBox.length.z = bbLengthZ;
-                }
-
-                if (paintStruct.contains("supportsHeightId"))
-                {
-                    const auto& id = paintStruct["supportsHeightId"];
-                    if (_heightMapping.find(id) != _heightMapping.end())
-                    {
-                        paint.HeightSupports = &_heightMapping[id];
-                    }
-                }
+                PaintStructJson paintJson(*this);
+                paintJson.FromJson(paintStruct);
+                paintJson.Key.FromJson(paintStruct);
 
                 // we need to check if the key exists
                 //_keys.push_back(key);
-                keysJson.push_back(keyJson);
-                _paintStructs.push_back(paint);
+                keysJson.push_back(paintJson.Key);
+                _paintStructs.push_back(paintJson.Value);
             }
 
             //auto keyGen = PaintStructKeyGenerator(keysJson);
